requirements.txt
# requirements.txt (Final, Hardened for Deployment)

# --- Core Frameworks ---
streamlit>=1.33.0
langchain>=0.1.16
langchain-core>=0.1.45
langchain-community>=0.0.34

# --- LLM & Embeddings ---
langchain-google-genai>=1.0.1
google-generativeai>=0.4.1

# --- Core Scientific & ML Libraries (Specify versions to ensure compatibility) ---
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers>=2.2.2

# --- Data Sources & Processing ---
pypdf>=4.0.0
beautifulsoup4>=4.12.3

# --- Databases (Vector & Graph) ---
faiss-cpu>=1.7.4

# --- ArangoDB Packages ---
python-arango>=7.8.0
langchain-arangodb>=0.1.0

# --- Web Search & Visualization ---
langchain-community>=0.0.34 # Already included, but for clarity
streamlit-agraph>=0.0.38

# --- Utilities ---
pydantic>=2.0.0
python-dotenv>=1.0.0